Make ts3.1 the default
