--- types/jasmine/ts3.1/tsconfig.json
@@ -1,22 +1,0 @@
-{
-    "files": [
-        "index.d.ts",
-        "jasmine-tests.ts"
-    ],
-    "compilerOptions": {
-        "module": "commonjs",
-        "lib": [
-            "es6",
-            "dom"
-        ],
-        "noImplicitAny": true,
-        "noImplicitThis": true,
-        "strictNullChecks": true,
-        "strictFunctionTypes": false,
-        "baseUrl": "../../",
-        "typeRoots": ["../../"],
-        "types": [],
-        "noEmit": true,
-        "forceConsistentCasingInFileNames": true
-    }
-}
