--- types/jasmine/ts3.1/tslint.json
@@ -1,12 +1,0 @@
-{
-    "extends": "dtslint/dt.json",
-    "rules": {
-        "ban-types": false,
-        "no-empty-interface": false,
-        "no-single-declare-module": false,
-        "no-unnecessary-generics": false,
-        "no-var-keyword": false,
-        "one-variable-per-declaration": false,
-        "only-arrow-functions": false
-    }
-}
