+++ types/jasmine/tslint.json
@@ -1,9 +1,7 @@
 {
     "extends": "dtslint/dt.json",
     "rules": {
-        "await-promise": false,
         "ban-types": false,
-        "no-declare-current-package": false,
         "no-empty-interface": false,
         "no-single-declare-module": false,
         "no-unnecessary-generics": false,
